+++ quaternion/src/quaternion.rs
@@ -7,7 +7,7 @@ use libm::{acos, sin};
 
 #[derive(Debug)]
 pub struct Quaternion {
-    vect: Vector<f64>,
+    vect: Vector<f32>,
 }
 
 impl Quaternion {
@@ -23,7 +23,7 @@ impl Quaternion {
     }
 
     // This function is used to fill a quaternion.
-    pub fn fill(self: &mut Self, qw: f64, qx: f64, qy: f64, qz: f64) -> Result<(), LinalgError> {
+    pub fn fill(self: &mut Self, qw: f32, qx: f32, qy: f32, qz: f32) -> Result<(), LinalgError> {
         self.vect.set_element(0, qw)?;
         self.vect.set_element(1, qx)?;
         self.vect.set_element(2, qy)?;
@@ -69,22 +69,22 @@ impl Quaternion {
     // This function is used to perform the multiplication operation (Hamilton product) of two quaternions of size (4 x 1).
     pub fn mul(&mut self, quat1: &Self, quat2: &Self) -> Result<(), LinalgError> {
         // Extract components of the first quaternion (quat1).
-        let qw1: f64 = quat1.vect.get_element(0)?;
-        let qx1: f64 = quat1.vect.get_element(1)?;
-        let qy1: f64 = quat1.vect.get_element(2)?;
-        let qz1: f64 = quat1.vect.get_element(3)?;
+        let qw1: f32 = quat1.vect.get_element(0)?;
+        let qx1: f32 = quat1.vect.get_element(1)?;
+        let qy1: f32 = quat1.vect.get_element(2)?;
+        let qz1: f32 = quat1.vect.get_element(3)?;
 
         // Extract components of the second quaternion (quat2).
-        let qw2: f64 = quat2.vect.get_element(0)?;
-        let qx2: f64 = quat2.vect.get_element(1)?;
-        let qy2: f64 = quat2.vect.get_element(2)?;
-        let qz2: f64 = quat2.vect.get_element(3)?;
+        let qw2: f32 = quat2.vect.get_element(0)?;
+        let qx2: f32 = quat2.vect.get_element(1)?;
+        let qy2: f32 = quat2.vect.get_element(2)?;
+        let qz2: f32 = quat2.vect.get_element(3)?;
 
         // Calculate the components of the resulting quaternion (self).
-        let qw: f64 = qw1 * qw2 - qx1 * qx2 - qy1 * qy2 - qz1 * qz2;
-        let qx: f64  = qw1 * qx2 + qx1 * qw2 - qz1 * qy2 + qy1 * qz2;
-        let qy: f64  = qw1 * qy2 + qz1 * qx2 + qw2 * qy1 - qx1 * qz2;
-        let qz: f64  = qw1 * qz2 - qy1 * qx2 + qx1 * qy2 + qw2 * qz1;
+        let qw: f32 = qw1 * qw2 - qx1 * qx2 - qy1 * qy2 - qz1 * qz2;
+        let qx: f32  = qw1 * qx2 + qx1 * qw2 - qz1 * qy2 + qy1 * qz2;
+        let qy: f32  = qw1 * qy2 + qz1 * qx2 + qw2 * qy1 - qx1 * qz2;
+        let qz: f32  = qw1 * qz2 - qy1 * qx2 + qx1 * qy2 + qw2 * qz1;
 
         // Store the components of in the resulting quaternion.
         self.fill(qw, qx, qy, qz)?;
@@ -96,22 +96,22 @@ impl Quaternion {
     // This function is used to perform the multiplication operation (Hamilton product) of two quaternions of size (4 x 1).
     pub fn mul(&mut self, p: &Self, q: &Self) -> Result<(), LinalgError> {
         // Extract components of the first quaternion (p).
-        let pw: f64 = p.vect.get_element(0)?;
-        let px: f64 = p.vect.get_element(1)?;
-        let py: f64 = p.vect.get_element(2)?;
-        let pz: f64 = p.vect.get_element(3)?;
+        let pw: f32 = p.vect.get_element(0)?;
+        let px: f32 = p.vect.get_element(1)?;
+        let py: f32 = p.vect.get_element(2)?;
+        let pz: f32 = p.vect.get_element(3)?;
 
         // Extract components of the second quaternion (q).
-        let qw: f64 = q.vect.get_element(0)?;
-        let qx: f64 = q.vect.get_element(1)?;
-        let qy: f64 = q.vect.get_element(2)?;
-        let qz: f64 = q.vect.get_element(3)?;
+        let qw: f32 = q.vect.get_element(0)?;
+        let qx: f32 = q.vect.get_element(1)?;
+        let qy: f32 = q.vect.get_element(2)?;
+        let qz: f32 = q.vect.get_element(3)?;
 
         // Calculate the components of the resulting quaternion (self).
-        let w: f64 = pw*qw - px*qx - py*qy - pz*qz;
-        let x: f64  = pw*qx + px*qw + py*qz - pz*qy;
-        let y: f64  = pw*qy - px*qz + py*qw + pz*qx;
-        let z: f64  = pw*qz + px*qy - py*qx + pz*qw;
+        let w: f32 = pw*qw - px*qx - py*qy - pz*qz;
+        let x: f32  = pw*qx + px*qw + py*qz - pz*qy;
+        let y: f32  = pw*qy - px*qz + py*qw + pz*qx;
+        let z: f32  = pw*qz + px*qy - py*qx + pz*qw;
 
         // Store the components of in the resulting quaternion.
         self.fill(w, x, y, z)?;
@@ -136,7 +136,7 @@ impl Quaternion {
 
     // This function is used to invert a quaternion.
     pub fn invert(self: &mut Self) -> Result<(), LinalgError> {
-        let norm: f64 = self.vect.calculate_norm()?;    // Calculate the norm of the quaternion.
+        let norm: f32 = self.vect.calculate_norm()?;    // Calculate the norm of the quaternion.
         self.conjugate()?;  // Conjugate the quaternion.
 
         if norm > EPSILON {
@@ -149,15 +149,15 @@ impl Quaternion {
     }
 
     // This function is used to multiply by a scalar all elements of a quaternion.
-    pub fn mul_by_scalar(self: &mut Self, scalar: f64) -> Result<(), LinalgError> {
+    pub fn mul_by_scalar(self: &mut Self, scalar: f32) -> Result<(), LinalgError> {
         self.vect.mul_by_scalar(scalar)?;
 
         Ok(())  // Return no error.
     }
 
     // This function is used to extract the vector from the quaternion object (can be used to convert a quaternion to a vector).
-    pub fn get_vect(self: &Self) -> Result<Vector<f64>, LinalgError> {
-        let mut vect: Vector<f64> = Vector::new();
+    pub fn get_vect(self: &Self) -> Result<Vector<f32>, LinalgError> {
+        let mut vect: Vector<f32> = Vector::new();
         vect.init(4)?;
         vect.copy_from(&self.vect)?;
 
@@ -165,10 +165,10 @@ impl Quaternion {
     }
 
     // This function is used to perform the spherical interpolation (SLERP) of two quaternions.
-    pub fn slerp(self: &mut Self, quat1: &Self, quat2: &Self, alpha: f64) -> Result<(), LinalgError> {
+    pub fn slerp(self: &mut Self, quat1: &Self, quat2: &Self, alpha: f32) -> Result<(), LinalgError> {
         // Perform scalar product.
-        let scalar: f64 = dot_product(&quat1.get_vect()?, &quat2.get_vect()?)?;
-        let theta: f64 = acos(scalar);  // Angle between the axis of the first quaternion and the second one.
+        let scalar: f32 = dot_product(&quat1.get_vect()?, &quat2.get_vect()?)?;
+        let theta: f32 = acos(scalar);  // Angle between the axis of the first quaternion and the second one.
 
         // Create a copy of the first quaternion.
         let mut q1: Self = Self::new()?;
@@ -178,7 +178,7 @@ impl Quaternion {
         let mut q2: Self = Self::new()?;
         q2.copy_from(&quat2)?;
 
-        q1.mul_by_scalar(sin((1.0_f64 - alpha)*theta) / sin(theta))?;
+        q1.mul_by_scalar(sin((1.0_f32 - alpha)*theta) / sin(theta))?;
         q2.mul_by_scalar(sin(alpha*theta) / sin(theta))?;
         self.add(&q1, &q2)?;
 
@@ -186,22 +186,22 @@ impl Quaternion {
     }
 
     // This function is used to obtain the qw value of a quaternion.
-    pub fn get_qw(self: &Self) -> Result<f64, LinalgError> {
+    pub fn get_qw(self: &Self) -> Result<f32, LinalgError> {
         Ok(self.vect.get_element(0)?)
     }
 
     // This function is used to obtain the qx value of a quaternion.
-    pub fn get_qx(self: &Self) -> Result<f64, LinalgError> {
+    pub fn get_qx(self: &Self) -> Result<f32, LinalgError> {
         Ok(self.vect.get_element(1)?)
     }
 
     // This function is used to obtain the qy value of a quaternion.
-    pub fn get_qy(self: &Self) -> Result<f64, LinalgError> {
+    pub fn get_qy(self: &Self) -> Result<f32, LinalgError> {
         Ok(self.vect.get_element(2)?)
     }
 
     // This function is used to obtain the qz value of a quaternion.
-    pub fn get_qz(self: &Self) -> Result<f64, LinalgError> {
+    pub fn get_qz(self: &Self) -> Result<f32, LinalgError> {
         Ok(self.vect.get_element(3)?)
     }
 
@@ -214,10 +214,10 @@ impl Quaternion {
 }
 
 // This function is used to perform the spherical interpolation (SLERP) of two quaternions.
-pub fn slerp(quat1: &Quaternion, quat2: &Quaternion, alpha: f64) -> Result<Quaternion, LinalgError> {
+pub fn slerp(quat1: &Quaternion, quat2: &Quaternion, alpha: f32) -> Result<Quaternion, LinalgError> {
     // Perform scalar product.
-    let scalar: f64 = dot_product(&quat1.get_vect()?, &quat2.get_vect()?)?;
-    let theta: f64 = acos(scalar);  // Angle between the axis of the first quaternion and the second one.
+    let scalar: f32 = dot_product(&quat1.get_vect()?, &quat2.get_vect()?)?;
+    let theta: f32 = acos(scalar);  // Angle between the axis of the first quaternion and the second one.
 
     // Create a copy of the first quaternion.
     let mut q1: Quaternion = Quaternion::new()?;
@@ -227,7 +227,7 @@ pub fn slerp(quat1: &Quaternion, quat2: &Quaternion, alpha: f64) -> Result<Quate
     let mut q2: Quaternion = Quaternion::new()?;
     q2.copy_from(&quat2)?;
 
-    q1.mul_by_scalar(sin((1.0_f64 - alpha)*theta) / sin(theta))?;
+    q1.mul_by_scalar(sin((1.0_f32 - alpha)*theta) / sin(theta))?;
     q2.mul_by_scalar(sin(alpha*theta) / sin(theta))?;
 
     let mut q: Quaternion = Quaternion::new()?;
